--- src/main/scala/za/co/absa/hyperdrive/trigger/persistance/SearchableRepository.scala
@@ -1,55 +1,0 @@
-
-/*
- * Copyright 2018 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.hyperdrive.trigger.persistance
-
-import slick.lifted.{AbstractTable, Query, TableQuery}
-import za.co.absa.hyperdrive.trigger.models.search.{TableSearchRequest, TableSearchResponse}
-import za.co.absa.hyperdrive.trigger.models.tables.SearchableTable
-
-import scala.concurrent.{ExecutionContext, Future}
-
-trait SearchableRepository extends Repository {
-  def search[T <: AbstractTable[_] with SearchableTable](tableQuery: TableQuery[T], request: TableSearchRequest)
-                                                        (implicit ec: ExecutionContext): Future[TableSearchResponse[T#TableElementType]] = {
-    val initQuery: Query[T, T#TableElementType, Seq] = tableQuery
-    val withStringEquals = request.getStringEqualsFilterAttributes.foldLeft(initQuery)((query, attributes) =>
-      query.filter(table => table.applyStringEqualsFilter(attributes)))
-    val withContains = request.getContainsFilterAttributes.foldLeft(withStringEquals)((query, attributes) =>
-      query.filter(table => table.applyContainsFilter(attributes)))
-    val withIntRange = request.getIntRangeFilterAttributes.foldLeft(withContains)((query, attributes) =>
-      query.filter(table => table.applyIntRangeFilter(attributes)))
-    val filteredQuery = request.getDateTimeRangeFilterAttributes.foldLeft(withIntRange)((query, attributes) =>
-      query.filter(table => table.applyDateTimeRangeFilter(attributes)))
-
-    import profile.api._
-    val length = filteredQuery.length.result
-    val result = filteredQuery
-      .sortBy(_.sortFields(request.sort))
-      .drop(request.from)
-      .take(request.size).result
-
-    db.run {
-        for {
-          l <- length
-          r <- result
-        } yield {
-          TableSearchResponse[T#TableElementType](items = r, total = l)
-        }
-    }
-  }
-
-}
